fixed profile picture adding and roomie name changing
