+++ Views/Shared/_Layout.cshtml
@@ -26,7 +26,7 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="IO" asp-action="RoomieProfile">Profile</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Profile" asp-action="RoomieProfile">Profile</a>
                         <li class="nav-item">
                             @{
                                 var userName = Context.Session.GetString("UserName");
